@@ -1,7 +1,7 @@
 <?php
-class PodioAppFieldCollectionTest extends PHPUnit_Framework_TestCase {
+class PodioAppFieldCollectionTest extends \PHPUnit\Framework\TestCase {
 
-  public function setup() {
+  public function setUp(): void {
     $this->collection = new PodioAppFieldCollection(array(
       new PodioAppField(array('field_id' => 1, 'external_id' => 'a', 'type' => 'text')),
       new PodioAppField(array('field_id' => 2, 'external_id' => 'b', 'type' => 'number')),
@@ -35,10 +35,8 @@ class PodioAppFieldCollectionTest extends PHPUnit_Framework_TestCase {
     $this->assertEquals($length+1, count($this->collection));
   }
 
-  /**
-    * @expectedException PodioDataIntegrityError
-    */
   public function test_cannot_add_item_field() {
+    $this->expectException('PodioDataIntegrityError');
     $this->collection[] = new PodioItemField();
   }
 
